+++ src/main/java/bowling/domain/rollresult/Gutter.java
@@ -1,4 +1,4 @@
-package bowling.rollresult;
+package bowling.domain.rollresult;
 
 import java.util.Objects;
