+++ src/test/java/bowling/domain/HitNumberTest.java
@@ -1,5 +1,6 @@
-package bowling;
+package bowling.domain;
 
+import bowling.domain.HitNumber;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
