+++ src/test/java/bowling/domain/PlayerTest.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 import org.junit.jupiter.api.Test;
