+++ src/test/java/bowling/domain/rollresult/MissTest.java
@@ -1,4 +1,4 @@
-package bowling.rollresult;
+package bowling.domain.rollresult;
 
 import org.junit.jupiter.api.Test;
