+++ src/test/java/bowling/domain/rollresult/OneRollResultTypeTest.java
@@ -1,4 +1,4 @@
-package bowling.rollresult;
+package bowling.domain.rollresult;
 
 import org.junit.jupiter.api.Test;
