+++ src/test/java/bowling/domain/rollresult/RollResultTest.java
@@ -1,7 +1,7 @@
-package bowling.rollresult;
+package bowling.domain.rollresult;
 
-import bowling.HitNumber;
-import bowling.Pin;
+import bowling.domain.HitNumber;
+import bowling.domain.Pin;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
